batch norm forward_t to be tested, added EPEPS (epsilon) macro

+++ src/core/BConvLayer.cpp
@@ -24,11 +24,11 @@ namespace bdlearn {
         return;
     }
 
-    uint8_t BConvLayer::get_w(size_t x, size_t y, size_t in_c, size_t out_c) {
+    uint8_t BConvLayer::get_w(int x, int y, int in_c, int out_c) {
         return w_.get(out_c, in_c * k_ * k_ + y * k_ + x);
     }
 
-    float BConvLayer::get_train_w(size_t x, size_t y, size_t in_c, size_t out_c) {
+    float BConvLayer::get_train_w(int x, int y, int in_c, int out_c) {
         return train_w_.get()[out_c * k_ * k_ * in_c_
                                 + in_c * k_ * k_
                                 + y * k_
