+++ test/src/BMat_basic.cpp
@@ -17,9 +17,9 @@ int test_BMat_copy_and_equals() {
 }
 
 int test_BMat_matmul_simple() {
-    size_t m = 3; 
-    size_t k = 7;
-    size_t n = 5;
+    int m = 3; 
+    int k = 7;
+    int n = 5;
     BMat s1(m, k);
     BMat s2(k, n);
     s1.random();
@@ -32,8 +32,8 @@ int test_BMat_matmul_simple() {
     float* disp = res;
     
     /*
-    for (size_t i = 0; i < m; ++i) {
-        for (size_t j = 0; j < n; ++j) {
+    for (int i = 0; i < m; ++i) {
+        for (int j = 0; j < n; ++j) {
             std::cout << *disp << " ";
             ++disp;
         }
@@ -41,10 +41,10 @@ int test_BMat_matmul_simple() {
     }
     */
 
-    for (size_t i = 0; i < m; ++i) {
-        for (size_t j = 0; j < n; ++j) {
+    for (int i = 0; i < m; ++i) {
+        for (int j = 0; j < n; ++j) {
             float sum = 0.0;
-            for (size_t x = 0; x < k; x++) {
+            for (int x = 0; x < k; x++) {
                 uint8_t t1 = s1.get(i, x);
                 uint8_t t2 = s2.get(x, j);
                 sum += (t1 ? 1 : -1) * (t2 ? 1 : -1);
@@ -62,13 +62,13 @@ int test_BMat_matmul_simple() {
 }
 
 int test_BMat_sign_constructor() {
-    size_t m = 17;
-    size_t n = 7;
+    int m = 17;
+    int n = 7;
     float rand[m*n];
-    for (size_t i = 0; i < m*n; ++i) rand[i] = std::rand();
+    for (int i = 0; i < m*n; ++i) rand[i] = std::rand();
     BMat test(m, n, rand);
-    for (size_t i = 0; i < m; ++i) {
-        for (size_t j = 0; j < n; ++j) {
+    for (int i = 0; i < m; ++i) {
+        for (int j = 0; j < n; ++j) {
             if (rand[i*n + j] >= 0 != test.get(i, j)) {
                 std::cerr << "test_BMat_sign_constructor failed at " << i << ", " << j;
                 std::cerr << ". Expected: " << (rand[i*n + j] >= 0) << ", got: " << test.get(i, j) << std::endl;
